\name{Bayesian Online ChangePoint Detection}
\alias{bocpdr}
\docType{package}

\title{
  Bayesian Online ChangePoint Detection
}

\description{
  Changepoint detection and so on.
}

\details{
  \tabular{ll}{
    Package: \tab bocpdr\cr
    Type: \tab Package\cr
    Version: \tab 0.1.1\cr
    Date: \tab 2017-05-17\cr
    License: \tab GPL-2\cr
  }
  ~~ An overview of how to use the package, including the most important ~~
  ~~ functions ~~
}

\author{
  Jani Anttila
  Markku Koskinen
  Maintainer: Jani Anttila <jani.v.anttil@gmail.com>
}

\references{

  Perälä, T., Kuparinen, A. 2015. Detecting regime shifts in fish stock dynamics.
  Canadian Journal of Fisheries and Aquatic Sciences. 10.1139/cjfas-2014-040

  Adams, R. P., MacKay, D. J. C, 2007. Bayesian online changepoint detection.
  arxiv.org/abs/0710.3742
}

\keyword{ package }

\seealso{
  ~~ Optional links to other man pages, e.g. ~~
  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

\examples{
  a <- 5
}
